+++ Examples/iOSExample/ViewController.swift
@@ -35,7 +35,7 @@ class ViewController: UIViewController, UICollectionViewDataSource {
   }()
 
   private lazy var gradients: [(String, UIColor)] = {
-    return [UIColor(hex: 0x3498db), UIColor(hex: 0xe74c3c)].colors(amount: 12).map { ($0.toHexString(), $0) }
+    return [UIColor(hex: 0x3498db), UIColor(hex: 0xe74c3c)].gradient.colorPalette(amount: 12).map { ($0.toHexString(), $0) }
   }()
 
   override func viewDidLoad() {