+++ Sources/Utils.swift
@@ -24,8 +24,6 @@
  *
  */
 
-import Foundation
-
 #if os(iOS) || os(tvOS) || os(watchOS)
   import UIKit
 #elseif os(OSX)